Add Handler support.

+++ src/main/java/org/alternadev/whatsup/IncompleteMessageException.java
@@ -6,18 +6,18 @@ public class IncompleteMessageException extends Exception {
      *
      */
     private static final long serialVersionUID = 42L * 63L;
-    private int[] data;
+    private byte[] data;
 
-    public IncompleteMessageException(String message, int[] input) {
+    public IncompleteMessageException(String message, byte[] input) {
         super(message);
         this.data = input;
     }
 
-    public int[] getData() {
+    public byte[] getData() {
         return data;
     }
 
-    public void setData(int[] data) {
+    public void setData(byte[] data) {
         this.data = data;
     }
 }
